make cubemap material builtin
